Move away from explicit substitutions
I am now following a similar presentation to:
https://www.cse.chalmers.se/~nad/publications/danielsson-types2006.pdf
Which has the nice advantage of only contexts need to be HITs.
This makes pattern matching on the original datatypes much more
feasible, instead of requiring use of an explicit eliminator, and just
generally means terms are closer to a normal form which will be nice
when implementing comparing types/terms.

I think I also made a neat observation in that just a substitution from
an empty context to another empty context and _↑_ is enough to define
build up all identity substitutions (i.e: identity substitutions do not
need to themselves be a constructor).