Implement substitution fully
Unfortunately Agda cannot see that the definition is terminating. It's
not a traditional case of mutually recursive calls without decreasing
argument size, but instead it points to function calls inside the type
signatures of some of the substitution helpers, which I am not sure how
to fix (hence {-# TERMINATING #-}).

Application also looks like it will be a little tricky to add to the
syntax. I think I will start with defining normal forms (i.e: to rule
out app (lam M) N cases.