Start work on syntax with explicit substitutions
Defining the syntax of normal forms with recursive substitutions is
hard (possibly impossible). After a lot of experimentation, I have
found a new nice point in the design space (explicit substitutions, but
only in terms)
I am currently defining normalisation with this syntax, and I don't
want to jinx it, but it is looking like it might work!