Add remaining substitution equations
It turns application is a little trickier than I anticipated.
I am pretty sure proving substitutions can be moved inside applications
is possible, but relies on a proof that no variables with pi-types
exist in an empty context, which is pretty heavy.

I think going with categorical application (even though it is more
unintuitive IMO) is probably the best solution here.