Clean up substitutions
Substitutions for terms in normal form are awkward because of the case
where an LHS could turn into a lambda and we need to reduce the redex.
Therefore, I have defined these substitutions over just a predicate on
terms that gets rid of all the explicit substitutions, which is at
least a step towards normalisation.