Split off non-confluent rewrites
We want to apply rewrites to equalities so they reduce and we don't
need to sprinkle all the proofs with '=[]=-uip'.
Unfortunately, a few of the rewrites we need introduce non-confluence
(at least in Agda's implementation). IMO they are still worth
including for the convenience and their actual danger is quite
limited (as they only rewrite equality proofs, which with UIP can
be shown to all be 'refl' anyway) but I have split them into a
separate module regardless for clarity.